Refactor Anki tools into domain modules
