+++ tests/test_cards_info.py
@@ -90,7 +90,8 @@ ROOT = Path(__file__).resolve().parent.parent
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
-from server import CardInfo, CardsInfoArgs, cards_info
+from anki_mcp import CardInfo, CardsInfoArgs
+from anki_mcp.tools.notes import cards_info
 
 
 def test_cards_info_normalizes_payload(monkeypatch):
