+++ tests/test_deck_tools.py
@@ -88,14 +88,8 @@ ROOT = Path(__file__).resolve().parent.parent
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
-from server import (
-    DeleteDecksArgs,
-    DeckInfo,
-    RenameDeckArgs,
-    delete_decks,
-    list_decks,
-    rename_deck,
-)
+from anki_mcp import DeleteDecksArgs, DeckInfo, RenameDeckArgs
+from anki_mcp.tools.decks import delete_decks, list_decks, rename_deck
 
 
 def _unwrap_tool(func):
