+++ tests/test_dedup_key.py
@@ -36,7 +36,8 @@ ROOT = Path(__file__).resolve().parent.parent
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
-from server import AddNotesArgs, NoteInput, add_from_model, add_notes
+from anki_mcp import AddNotesArgs, NoteInput
+from anki_mcp.tools.notes import add_from_model, add_notes
 
 
 @pytest.fixture
