+++ tests/test_delete_notes.py
@@ -40,7 +40,8 @@ ROOT = Path(__file__).resolve().parent.parent
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
-from server import DeleteNotesArgs, delete_notes  # noqa: E402  # pylint: disable=wrong-import-position
+from anki_mcp import DeleteNotesArgs  # noqa: E402  # pylint: disable=wrong-import-position
+from anki_mcp.tools.notes import delete_notes  # noqa: E402  # pylint: disable=wrong-import-position
 
 
 @pytest.fixture
