+++ tests/test_media_tools.py
@@ -42,7 +42,8 @@ ROOT = Path(__file__).resolve().parent.parent
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
-from server import DeleteMediaArgs, MediaRequest, delete_media, get_media  # noqa: E402
+from anki_mcp import DeleteMediaArgs, MediaRequest  # noqa: E402
+from anki_mcp.tools.media import delete_media, get_media  # noqa: E402
 
 
 def _unwrap(tool):
