+++ tests/test_sync.py
@@ -40,7 +40,7 @@ ROOT = Path(__file__).resolve().parent.parent
 if str(ROOT) not in sys.path:
     sys.path.insert(0, str(ROOT))
 
-from server import sync  # noqa: E402
+from anki_mcp.tools.sync import sync  # noqa: E402
 
 
 def _unwrap(tool):
